ssl filter goes in its own file(s)
